refactor: 테스트 model 패키지 구성
