+++ src/test/java/lotto/model/LottoTest.java
@@ -1,10 +1,9 @@
-package lotto;
+package lotto.model;
 
 import static lotto.util.ErrorMessage.ERROR_LOTTO_OUT_OF_RANGE;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
-import lotto.model.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
